backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
dnspython==2.4.2
python-whois==0.8.0
shodan==1.31.0
openai==1.6.1
beautifulsoup4==4.12.2
lxml==4.9.3
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2